Code/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.2.0=pyhd8ed1ab_0
aiohttp=3.8.1=py310h5764c6d_1
aiosignal=1.2.0=pyhd8ed1ab_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.1.0=pyh71513ae_1
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=pypi_0
blas=1.0=mkl
bleach=5.0.1=pypi_0
blinker=1.4=py_1
bottleneck=1.3.5=py310ha9d4c09_0
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.6.15=ha878542_0
cachetools=5.2.0=pyhd8ed1ab_0
certifi=2022.6.15=py310hff52083_0
cffi=1.15.1=py310h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=py310hff52083_0
cryptography=37.0.1=py310h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
debugpy=1.5.1=py310h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pypi_0
dill=0.3.5.1=pypi_0
diskcache=5.4.0=pypi_0
entrypoints=0.4=py310h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
fastjsonschema=2.16.1=pypi_0
fastprogress=1.0.3=pypi_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.8.0=pypi_0
fonttools=4.37.1=pypi_0
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=py310h7f8727e_1
gensim=4.2.0=pypi_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.10.0=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
grpcio=1.42.0=py310hce63b2e_0
huggingface-hub=0.8.1=pypi_0
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.4=py310hff52083_0
importlib-resources=5.9.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.9.1=py310h06a4308_0
ipython=8.4.0=py310h06a4308_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.7.1=pypi_0
jedi=0.18.1=py310h06a4308_1
jinja2=3.1.2=pypi_0
joblib=1.1.0=pypi_0
jpeg=9e=h7f8727e_0
jsonschema=4.9.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-console=6.4.4=pypi_0
jupyter_client=7.2.2=py310h06a4308_0
jupyter_core=4.10.0=py310h06a4308_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=1.1.1=pypi_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libidn2=2.3.2=h7f8727e_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.19.1=h4ff587b_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h2818925_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libvpx=1.7.0=h439df22_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
lxml=4.9.1=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.4.1=pyhd8ed1ab_0
markupsafe=2.1.1=py310h5764c6d_1
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
multidict=6.0.2=py310h5764c6d_1
nbclient=0.6.6=pypi_0
nbconvert=6.5.3=pypi_0
nbformat=5.4.0=pypi_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py310h06a4308_0
nettle=3.7.3=hbbd107a_1
notebook=6.4.12=pypi_0
numexpr=2.8.3=py310hcea2de6_0
numpy=1.23.1=py310h1794996_0
numpy-base=1.23.1=py310hcba007f_0
oauthlib=3.2.0=pyhd8ed1ab_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1q=h7f8727e_0
paccmann-predictor=0.0.4=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.3=py310h6a678d5_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py310hace64e9_1
pip=22.1.2=py310h06a4308_0
prometheus-client=0.14.1=pypi_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.19.1=py310h295c915_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pubchempy=1.0.4=pypi_0
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd3eb1b0_0
pyfaidx=0.7.1=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.4.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.18.1=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.4=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.10=2_cp310
pytoda=1.0.2=pypi_0
pytorch=1.12.1=py3.10_cuda11.3_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py310h06a4308_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0=pypi_0
pyzmq=23.2.0=py310h6a678d5_0
qtconsole=5.3.1=pypi_0
qtpy=2.2.0=pypi_0
rdkit-pypi=2022.3.5=pypi_0
readline=8.1.2=h7f8727e_1
regex=2022.7.25=pypi_0
requests=2.28.1=py310h06a4308_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scikit-learn=1.1.2=pypi_0
scipy=1.9.0=pypi_0
seaborn=0.11.2=pypi_0
selfies=2.1.1=pypi_0
send2trash=1.8.0=pypi_0
setuptools=61.2.0=py310h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smart-open=6.1.0=pypi_0
smilespe=0.0.3=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.39.2=h5082296_0
stack_data=0.2.0=pyhd3eb1b0_0
tensorboard=2.10.0=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py310h597c629_2
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
terminado=0.15.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tinycss2=1.1.1=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.12.1=pypi_0
torch-tb-profiler=0.4.0=pypi_0
torchaudio=0.12.1=py310_cu113
torchvision=0.13.1=py310_cu113
tornado=6.1=py310h7f8727e_0
tqdm=4.64.0=pypi_0
traitlets=5.3.0=pypi_0
transformers=4.21.1=pypi_0
typing-extensions=4.3.0=py310h06a4308_0
typing_extensions=4.3.0=py310h06a4308_0
tzdata=2022a=hda174b7_0
unidecode=1.3.4=pypi_0
upfp=0.0.5=pypi_0
urllib3=1.26.11=py310h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=pypi_0
werkzeug=2.2.2=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.6.1=pypi_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.5=h7f8727e_1
yarl=1.7.2=py310h5764c6d_2
zeromq=4.3.4=h2531618_0
zipp=3.8.1=pyhd8ed1ab_0
zlib=1.2.12=h7f8727e_2
zstd=1.5.2=ha4553b6_0